feat: cleaned code, moved redux hook -> utils
